+++ core/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to the `@autonomys/agent-core` package will be documented in
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - 2025-05-10
+
+### Fixed
+- Bug with MCP return types - make it compatible with new tool execution flow
+
+### Changed
+- Update barrel files of chat agent
+
+### Added
+- langchain deepseek pacakge
+
 ## [0.3.0] - 2025-05-08
 
 ### Added